Switch to maven
Currently the fatJar task does not work correctly and includes test dependencies.
The easy solution is to not use the fatJar task, but because there's no pom file
there's no way to resolve the transitive dependencies.

As such, it's easier just to convert this to maven rather than trying to
fix the fatJar task, which really is kind of a hack.  I think the ideal
solution is to use the shadow plugin.